# add secretGenerator