@mcollina 

Is this what you had in mind with using randomInt?

I dont even know if it is really async how i implemented .secret().  I could implement the async secret() with randomBytes() and only secretSync using the randomInt secretGenerator.



before:
```
> /home/uzlopak/.nvm/versions/node/v16.14.2/bin/node benchmark/secret.js

  secret

  3 tests completed.

  secretSync        x 555,162 ops/sec ±0.50% (190 runs sampled)
  secret - callback x 113,889 ops/sec ±1.23% (172 runs sampled)
  secret - promise  x 109,406 ops/sec ±1.19% (176 runs sampled)
```

after:
```
uzlopak@Battlestation:~/workspace/csrf$ node benchmark/secret.js 

  secret

  3 tests completed.

  secretSync        x 852,334 ops/sec ±0.27% (193 runs sampled)
  secret - callback x 147,124 ops/sec ±1.49% (128 runs sampled)
  secret - promise  x 792,502 ops/sec ±0.57% (190 runs sampled)
```