@mcollina 2022-07-11
No. I mean to update nodejs/node to actuall cache randomBytes too.